requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.5.6
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.5.7
better-profanity==0.6.1
black==22.6.0
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
discord.py @ git+https://github.com/Rapptz/discord.py@44abc469f19cb03c35f10d8f41a10ca954e922f4
dnspython==1.16.0
flake8==3.9.2
frozenlist==1.3.1
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.5.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mongomock==4.1.2
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prettier==0.0.7
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pymongo==4.2.0
PyNaCl==1.3.0
pyparsing==3.0.9
python-dotenv==0.18.0
sentinels==1.0.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.6.16.1
zipp==3.4.1
